modules/cc_setup/python_src/bq_src/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-auth